+++ spec/features/checklist_resolver_spec.rb
@@ -3,11 +3,13 @@ describe "features" do
     %i(all_fields sciname sciname_auth sciname_rank).each do |input|
       context input do
         it "resolves #{input}" do
-          output = "/tmp/#{input}-processed.csv"
-          input = FILES[input]
-          FileUtils.rm(output) if File.exist?(output)
-          GnCrossmap.run(input, output, 1, true)
-          expect(File.exist?(output)).to be true
+          opts = { output: "/tmp/#{input}-processed.csv",
+                   input: FILES[input],
+                   data_source_id: 1,
+                   skip_original: true }
+          FileUtils.rm(opts[:output]) if File.exist?(opts[:output])
+          GnCrossmap.run(opts)
+          expect(File.exist?(opts[:output])).to be true
         end
       end
     end
@@ -15,10 +17,11 @@ describe "features" do
 
   context "combining acceptedName output" do
     it "gives accepted name for all matches" do
-      output = "/tmp/output.csv"
-      input = FILES[:sciname]
-      GnCrossmap.run(input, output, 1, true)
-      CSV.open(output, col_sep: "\t", headers: true).each do |r|
+      opts = { output: "/tmp/output.csv",
+               input: FILES[:sciname],
+               data_source_id: 1, skip_original: true }
+      GnCrossmap.run(opts)
+      CSV.open(opts[:output], col_sep: "\t", headers: true).each do |r|
         next unless r["matchedEditDistance"] == "0"
         expect(r["matchedName"].size).to be > 1
         expect(r["acceptedName"].size).to be > 1
@@ -28,31 +31,33 @@ describe "features" do
           expect(r["matchedName"]).to eq r["acceptedName"]
         end
       end
-      FileUtils.rm(output)
+      FileUtils.rm(opts[:output])
     end
   end
 
   context "use alternative headers" do
     it "uses alternative headers for resolution" do
-      output = "/tmp/output.csv"
-      input = FILES[:no_taxonid]
-      alt_headers = %w(taxonID scientificName rank)
-      GnCrossmap.run(input, output, 1, true, alt_headers)
-      CSV.open(output, col_sep: "\t", headers: true).each do |r|
+      opts = { output: "/tmp/output.csv",
+               input: FILES[:no_taxonid],
+               data_source_id: 1, skip_original: true,
+               alt_headers: %w(taxonID scientificName rank) }
+      GnCrossmap.run(opts)
+      CSV.open(opts[:output], col_sep: "\t", headers: true).each do |r|
         next unless r["matchedEditDistance"] == "0"
         expect(r["matchedName"].size).to be > 1
         expect(r["acceptedName"].size).to be > 1
       end
-      FileUtils.rm(output)
+      FileUtils.rm(opts[:output])
     end
 
     it "breaks without alternative headers" do
-      output = "/tmp/output.csv"
-      input = FILES[:no_taxonid]
+      opts = { output: "/tmp/output.csv",
+               input: FILES[:no_taxonid],
+               data_source_id: 1, skip_original: 1 }
       expect do
-        GnCrossmap.run(input, output, 1, true)
+        GnCrossmap.run(opts)
       end.to raise_error GnCrossmapError
-      FileUtils.rm(output)
+      FileUtils.rm(opts[:output])
     end
   end
 end